requirements.txt
black==23.3.0
click==8.1.3
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
mysqlclient==2.1.1
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
python-dotenv==1.0.0
SQLAlchemy==2.0.9
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.2.3
